fix: specify integers in some structures

+++ src/lib/api/structures/guild/economy/item.ts
@@ -18,7 +18,7 @@ export const ItemStructure = z
 		description: z.string().min(1).max(1024).optional(),
 		price: z.number().int().min(0).max(Number.MAX_SAFE_INTEGER),
 		type: z.nativeEnum(ItemType),
-		stock: z.number().min(0).max(200_000).optional(),
+		stock: z.number().int().min(0).max(200_000).optional(),
 		available_until: PreprocessedDateStructure.optional(),
 		value: z.string().nullable()
 	})
